Beginning of out-of-band "is the server alive" checks.
The situations I'm worried about are:
  * the server gcdeaths and gets into a state where the connection is
    alive as far as the OS is concerned but the server is unresponsive.
  * the server completely drops off the network (OS locks up or the power
    cord is yanked) and so the client blocks writing its request.
In addition, httpclient provides no way to time out on request send.
You can time out on connect, and you can time out on response receive,
but not request send.

So what we'll do is set up a parallel UDP socket listening on the same
port as the HTTP server which answers UDP ping packets.  If we send
n in a row (where n == 5, currently) without getting a response, we'll
assume it's dead and abort the request.

We'll probably want to extend this ping check over receiving the response
too.

Currently the response "secret" is hardcoded.  We'll want to pass that in.